refactor: split player state module
